mv OpGraph into a seperate file op_graph.h
