+++ oneflow/core/operator/reduce_concat_op.cpp
@@ -19,7 +19,6 @@ const PbMessage& ReduceConcatOp::GetCustomizedConf() const {
 void ReduceConcatOp::InferBlobDescs(std::function<BlobDesc*(const std::string&)> GetBlobDesc4BnInOp,
                                     const ParallelContext* parallel_ctx) const {
   int32_t in_num = op_conf().reduce_concat_conf().in_num();
-  CHECK_GE(in_num, 2);
   BlobDesc* first_in_blob = GetBlobDesc4BnInOp(input_bns().Get(0));
   BlobDesc* out_blob = GetBlobDesc4BnInOp(SoleObn());
   *out_blob = *first_in_blob;
